fix infinite calls problem in filter status
